backend/requirements.txt
django>=5.0,<6.0
djangorestframework>=3.15
django-cors-headers>=4.3
python-dotenv>=1.0
requests>=2.31
sentence-transformers>=3.0
qdrant-client>=1.9
openai>=1.30
torch>=2.0
